fix bug in InMemory _retrieve
Summary: -end_time needs to be a number, not a unicode

Reviewers: usmanm, marcua

Reviewed By: usmanm

Differential Revision: http://phabricator.locu.com/D3034

+++ kronos/server.py
@@ -215,7 +215,7 @@ def get_events(environment, start_response, headers):
   backend, configuration = router.backend_to_retrieve(request_json['stream'])
   events_from_backend = backend.retrieve(request_json['stream'],
                                          request_json.get('start_time'),
-                                         request_json['end_time'],
+                                         long(request_json['end_time']),
                                          request_json.get('start_id'),
                                          request_json.get('order',
                                            ResultOrder.ASCENDING),
